Implement the bootinfo boot command
It's now possible to get info on where to get the kernel, initrd and
boot params through the webserver interface. This is needed when booting
PPC64 VMs on mama guests. The ip config on the kernel command line is
now removed and all machines are set up with network manager instead.

Signed-off-by: Patrik Jakobsson <pjakobsson@suse.de>